Bumped version to 0.2.6
Added a simple CLI for testing

Initial cut at HWI integratino for Series 4/8 P5 processors.
This is a WIP snapshot before cleaning up the code. Can
  - retrieve and parse the DB
  - get/change light levels
  - press keypad buttons/switch buttons
  - get notifications when a light value changes
  - works within Home Assistant

Note: Uses ftp to grab the XML DB. The processor doesn't
work in PASSV mode, so firewalls need to be open. In docker,
the big hammer approach is to run as --net=host